Chatbot/langchain-chatbot-multiple-PDF-master/WorkingCode_2909/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.0.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.8
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.202
langchainplus-sdk==0.0.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.16
streamlit==1.12.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.15.0

pandasai==0.2.15
pdfplumber
streamlit-chat==0.0.2.2
youtube-transcript-api==0.6.0
